requirements.txt
ipywidgets==7.7.0
matplotlib==3.5.2
notebook==6.4.11
numpy==1.21.5
pandas==1.3.5
torch==1.11.0
torchvision
torchaudio
pytorch_lightning==1.6.3
seaborn==0.11.2
tensorboard==2.9.0
torchmetrics==0.8.2
tqdm==4.64.0
wandb==0.12.16
ipyvtklink==0.2.2


nibabel==3.2.2
timm==0.5.4
protobuf==3.20
warmup_scheduler==0.3
pyvista==0.34.1


einops==0.4.1
evalutils==0.3.1
scikit-learn==0.24.2
scipy==1.6.3
trimesh==3.12.0
vit-pytorch==0.33.2
python-dotenv==0.20.0
